pyproject.toml
[project]
name = "Backend_Dev_Capstone_Meta"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "NicolasB8854", email = "nicolasbogun@gmail.com"},
]
dependencies = [
    "django>=5.0.7",
    "mysqlclient>=2.2.4",
    "djangorestframework>=3.15.2",
    "djoser>=2.2.3",
    "djangorestframework-simplejwt>=5.3.1",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
